Fixed animating waves on all devices, again
